- updated to SLikeNet 0.1.1

+++ DependentExtensions/Swig/MakeSwigWithExtras.bat
@@ -1,7 +1,13 @@
-REM This file was taken from RakNet 4.082 without any modifications.
-REM Please see licenses/RakNet license.txt for the underlying license and related copyright.
-
 @echo off
+
+REM This file was taken from RakNet 4.082.
+REM Please see licenses/RakNet license.txt for the underlying license and related copyright.
+REM
+REM Modified work: Copyright (c) 2018, SLikeSoft UG (haftungsbeschränkt)
+REM
+REM This source code was modified by SLikeSoft. Modifications are licensed under the MIT-style
+REM license found in the license.txt file in the root directory of this source tree.
+
 SetLocal EnableDelayedExpansion
 set sourceDir=%1
 CALL :dequote sourceDir
@@ -13,7 +19,7 @@ if "%swigPath%"=="" goto :SKIPADDSLASH
 if "%swigPath:~-1%"=="\" goto :SKIPADDSLASH 
 SET swigPath=%swigPath%\
 :SKIPADDSLASH
-del /F /Q SwigOutput\SwigCSharpOutput\*
+if exist SwigOutput\SwigCSharpOutput del /F /Q SwigOutput\SwigCSharpOutput\*
 set option=%4
 CALL :dequote option
 set dependentDir=%3
